cppad_py: test_gaussian_cdf: convert gaussian_cdf_cam.py to pytest format.
